netkvm: process NetPnp events and detect protocol bind completion
The best event to detect protocol bind completion is NetEventRestart.
Note that event NetEventBindsComplete seems more suitable but in
fact it is not, as it is not related to any specific binding and
does not come with protocol binding context pointer, it is related
to all the adapters.
When we detect competion of protocol binding, we schedule work
item to attach the VF adapter to the NETKVM instance (if present).

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>