netkvm: attach the VF to NETKVM instance only when binding is completed
https://bugzilla.redhat.com/show_bug.cgi?id=1912088
If we indicate a network change to obtain an IP address for
NETKVM adapter in SRIOV failover setup, it sometimes requires
a long time (30 sec and more) to receive an IP address.
According to the research it might happen because the DHCP
handshaking is broken by following pause-restart after the
VF adapter is bound to the protocol.
Now we attach the METKVM to VF and indicate the network change
after the protocol is bound to the VF and VF will not be paused
anymore.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>